Finished Minimum, Core, Completion
Reads all files (except polygons) and adds them into a data structure. I have also implemented the draw method which draws the nodes and segments. The onClick method highlights a node (if in the vicinity of the click location) and highlights it. Also the roads which are connected to it are also listed in the text output box. onSearch has also been implemented and a Trie structure of the road names allows for prefix search. I have also implemented mouseWheelListener which allows zooming to happen with the mouse wheel